UI - WebUI.java - New base interface for all fully-custom UIs within WebLaF - All fully-custom UIs now provide default font, background and foreground values into components [ #428 ]
Styling
- com.alee.managers.style.validation - New package containing all XSDs for skin XML files [ #439 ]
- Added reference to XML Schema Definition (XSD) within each separate skin file
- AbstractPainter.java, AbstractDecorationPainter.java, ... - Added defence against possible NPEs within listeners [ #442 ]
- PainterSupport.java - Refactored and some previously public internal methods are now private
- PainterStyleConverter.java - Added possibility to specify painter classes by their XStream aliases
- Sides.java, SidesConverter.java - Added for convenient sides declaration within code
- StyleException.java, ComponentStyleConverter.java - Refactored exceptions creation code

OptionPane
- WebOptionPaneUI.java - Overridden default content creation methods to avoid unnecessary operations [ #367 ]
- StyleId.java, optionpane.xml - Added more child styles for all existing components within option pane
- skin.xml, optionpane.xml - Added separate dark style

DesktopPane, InternalFrame
- WebInternalFrameTitlePane.java - Fixed double restore icons when maximized frame was iconified
- StyleId.java, desktoppane.xml - Added `transparent` style for convenience
- desktoppane.xml - Added appropriate default background settings
- internalframe.xml - Slightly refactored default style
- skin.xml, internalframe.xml - Added separate dark style

InterfaceInspector
- InterfaceTreeNode.java - Added fix to avoid events from firing twice on containers where component was added
- InterfaceTree.java, InterfaceTreeDataProvider.java - Added proper `ComponentHighlighter` filtering
- AbstractComponentInfo.java - Added small marker for opaque components to distinct them from non-opaque ones easily
- StyleableInfo.java - Added proper glass pane component check
- ComponentHighlighter.java - Made default scheme colors public

Popup
- PopupLayer.java - Fixed issue with popup layer becoming opaque on style change
- WebPopup.java - Added `draggable` option for convenience
- innerpopup.xml - Streamlined default inner popup style to match normal popup style
- skin.xml, innerpopup.xml - Added separate dark style

Tooltip
- HotkeyManager.java - Removed redundant custom label component
- customtooltip.xml - Fixed foreground color for tooltip text and hotkey labels [ #435 ]
- customtooltip.xml - Changed hotkey style to remove redundant elements
- skin.xml, customtooltip.xml - Added separate dark style

List
- WListUI.java, WebListUI.java - Added method for proper list layout update access
- WebFileList.java - Added calls to update list layout on preferred row or column count change [ #438 ]

Separator
- Stripes.java - More abstract replacement for older `SeparatorLines`, now extends `AbstractContent` and performs content painting
- Stripe.java - More abstract replacement for older `SeparatorLine`, still a simple data class
- SeparatorStripes.java - Renamed from `SeparatorContent` and simply extends `Stripes` content implementation
- LineContent.java - Removed as redundant duplicate

Menu
- menuitem.xml, checkboxmenuitem.xml, radiobuttonmenuitem.xml - Reduced accelerator preferred size to fit into menu item properly
- menu.xml, menuitem.xml, checkboxmenuitem.xml, radiobuttonmenuitem.xml - Streamlined visual padding
- menu.xml - Added hover state for menubar case
- radiobuttonmenuitem.xml - Fixed selected state style icon padding

Notification
- notification.xml - Added missing inner notification style

FileChooser
- filechooser.xml - Changed bottom toolbar gradient position for better visuals

Label, StyledLabel
- label.xml, styledlabel.xml - Removed redundant constraints

ProgressBar
- MovingHighlightBackground.java, progressbar.xml - Fixed typo in "passses" variable

PathField
- WebPathField.java - Removed unnecessary component menu setup

RootPane
- RootPanePainter.java - Root pane will now be opaque when no custom decoration provided
- rootpane.xml - Added default component background color to avoid default background color on natively-decorated windows
- rootpane.xml - Added missing settings in dark rootpane style

MenuGenerator
- JMenuBarGenerator.java, JMenuGenerator.java, JPopupMenuGenerator.java - New generators for common Swing menus
- MenuGenerator.java - Added constructors to match all available `WebMenu` constructors
- MenuGenerator.java - Menu language is now properly updated upon generator prefix change
- MenuBarGenerator.java, MenuGenerator.java, PopupMenuGenerator.java - Slightly changed flow within constructor calls

Utilities
- JarUtils.java - Separated from `ReflectUtils` as it provides a different set of utilities
- NonEventThreadHandler.java - Moved to a separate package along with its implementations
- ExceptionNonEventThreadHandler.java - Default implementation for `NonEventThreadHandler` that throws an exception
- PackageFilteringNonEventThreadHandler.java - Custom implementation for `NonEventThreadHandler` that helps to filter out exceptions from the same origin
- DebugUtils.java - Enhanced border debug information painting method
- LafUtils.java - Added methods to install and uninstall component defaults
- SwingUtils.java - Added method to retrieve focus accelerator key mask
- TextUtils.java - Added `arrayToString` methods allowing convenient array to string conversion
- FileUtils.java - Slightly enhanced `copyFile` methods to avoid possible NPEs

Demo
- WindowsGroup.java, JOptionPaneExample.java - Added `JOptionPane` usage examples
- DesktopPaneGroup.java, JInternalFrameExample.java - Added `JInternalFrame` usage example
- MenusGroup.java, JMenuBarExample.java - Added `JMenuBar` usage examples
- DemoApplication.java - Disabled north toolbar's possibility to detach from main window

General
- Fixed some old code formatting issues with annotations

+++ modules/core/src/com/alee/utils/FileUtils.java
@@ -1411,26 +1411,9 @@ public final class FileUtils
      */
     public static boolean copyFile ( final String src, final String dst )
     {
-        try
-        {
-            // Creating destination directory if needed
-            final File dstDir = new File ( new File ( dst ).getParent () );
-            if ( ensureDirectoryExists ( dstDir ) )
-            {
-                final FileChannel srcFC = new FileInputStream ( src ).getChannel ();
-                final FileChannel dstFC = new FileOutputStream ( dst ).getChannel ();
-                return copyFile ( srcFC, dstFC );
-            }
-            else
-            {
-                return false;
-            }
-
-        }
-        catch ( final FileNotFoundException e )
-        {
-            return false;
-        }
+        final File srcFile = src != null ? new File ( src ) : null;
+        final File dstFile = dst != null ? new File ( dst ) : null;
+        return copyFile ( srcFile, dstFile );
     }
 
     /**
@@ -1444,13 +1427,12 @@ public final class FileUtils
      */
     public static boolean copyFile ( final File srcFile, final File dstFile )
     {
-        if ( srcFile.exists () && srcFile.isFile () )
+        if ( srcFile != null && srcFile.exists () && srcFile.isFile () && dstFile != null )
         {
             try
             {
                 // Creating destination directory if needed
-                final File dstDir = new File ( dstFile.getParent () );
-                if ( ensureDirectoryExists ( dstDir ) )
+                if ( ensureDirectoryExists ( dstFile.getParentFile () ) )
                 {
                     final FileChannel srcFC = new FileInputStream ( srcFile ).getChannel ();
                     final FileChannel dstFC = new FileOutputStream ( dstFile ).getChannel ();
@@ -1936,11 +1918,7 @@ public final class FileUtils
             // Checking stop flag
             if ( listener != null && listener.shouldStopDownload () )
             {
-                out.flush ();
-                out.close ();
-                in.close ();
-                deleteFile ( dstFile );
-                return null;
+                return haltDownload ( dstFile, in, out );
             }
 
             // Downloading content part by part
@@ -1957,11 +1935,7 @@ public final class FileUtils
                     // Checking stop flag
                     if ( listener.shouldStopDownload () )
                     {
-                        out.flush ();
-                        out.close ();
-                        in.close ();
-                        deleteFile ( dstFile );
-                        return null;
+                        return haltDownload ( dstFile, in, out );
                     }
                 }
                 out.write ( buf, 0, bytesRead );
@@ -1986,6 +1960,24 @@ public final class FileUtils
             }
             return null;
         }
+    }
+
+    /**
+     * Halts download by properly closing streams and removing destination file.
+     *
+     * @param dstFile destination file
+     * @param in      input stream
+     * @param out     output stream
+     * @return always {@code null}
+     * @throws IOException if some IO operation failed
+     */
+    private static File haltDownload ( final File dstFile, final InputStream in, final FileOutputStream out ) throws IOException
+    {
+        out.flush ();
+        out.close ();
+        in.close ();
+        deleteFile ( dstFile );
+        return null;
     }
 
     /**
@@ -2205,7 +2197,7 @@ public final class FileUtils
      */
     public static boolean ensureDirectoryExists ( final File dir )
     {
-        return dir.exists () || dir.mkdirs ();
+        return dir != null && ( dir.exists () || dir.mkdirs () );
     }
 
     /**
